requirements.txt
click
pyyaml
requests
flask      # For web API (optional)
fastapi    # For web API (optional)
